Refactor user dialogs
